Pull member functions up to Interceptor
